Search for taxa using vernacular names

Search for taxa using vernacular names

fauna_by_vernacular(data, names, exact = FALSE)

Arguments

data

(data.frame) the data.frame imported with the load_faunabr function or generated with the function select_fauna.

names

(character) vernacular name ("Nome comum") of the species to be searched

exact

(logic) if TRUE, the function will search only for exact matches. For example, if names = "veado-mateiro" and exact = TRUE, the function will return only the species popularly known as "veado-mateiro". On the other hand, if names = "veado-mateiro" and exact = FALSE, the function will return other results as "Veado-mateiro-pequeno". Default = FALSE

Value

a data.frame with the species with vernacular names that match the input names

Examples

data("fauna_data") #Load Fauna do Brasil data
#Search for species whose vernacular name is 'veado-mateiro'
veado_exact <- fauna_by_vernacular(data = fauna_data,
                                   names = "veado-mateiro",
                                   exact = TRUE)
veado_exact
#>               species subspecies                  scientificName acceptedName
#> 2538 Mazama americana       <NA> Mazama americana Erxleben, 1777         <NA>
#> 7501      Mazama rufa       <NA>       Mazama rufa Illiger, 1815         <NA>
#>       kingdom   phylum    class        order   family  genus  lifeForm
#> 2538 Animalia Chordata Mammalia Artiodactyla Cervidae Mazama herbivore
#> 7501 Animalia Chordata Mammalia Artiodactyla Cervidae Mazama herbivore
#>          habitat                                    states
#> 2538 terrestrial                               AM;AP;PA;RR
#> 7501 terrestrial AC;AM;DF;GO;MA;MG;MS;MT;PR;RO;RS;SC;SP;TO
#>                     countryCode origin taxonomicStatus nomenclaturalStatus
#> 2538 BO;BR;CO;EC;GF;GY;SR;TT;VE native        accepted                <NA>
#> 7501                AR;BO;BR;PY native        accepted                <NA>
#>                                                                                                   vernacularName
#> 2538 Veado-mateiro, Veado-pardo, Corzuela colorada, Huazo, Soche colorado, Venado matacan rojo, Red Brocket deer
#> 7501                                                                                               Veado mateiro
#>      taxonRank     id language
#> 2538   species  50307       en
#> 7501   species 213525       en
#Search for species whose vernacular name is 'veado_mateiro', allowing non-exact
#matches
veado_not_exact <- fauna_by_vernacular(data = fauna_data,
                                       names = "veado-mateiro",
                                      exact = FALSE)